+++ roborockLib/lib/deviceFeatures.js
@@ -1412,5 +1412,6 @@ function supportsMaxPlusFanPower(model) {
 module.exports = {
   deviceFeatures,
   stateCodes,
+  errorCodes,
   supportsMaxPlusFanPower,
 };